+++ srcs/utils/init.c
@@ -1,0 +1,40 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   init.c                                             :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: adelille <marvin@42.fr>                    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2020/10/01 11:23:31 by adelille          #+#    #+#             */
+/*   Updated: 2021/12/02 22:55:36 by adelille         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "../../includes/bsq.h"
+
+void	init_data(t_data *data)
+{
+	data->nbr_lines = 0;
+	data->len_lines = 0;
+	data->stdin = NULL;
+}
+
+unsigned short	**init_map_file(t_data *d, unsigned short **map)
+{
+	int	i;
+
+	map = malloc(sizeof(unsigned short *) * d->nbr_lines);
+	if (!map)
+	{
+		ft_pser("Error: Malloc failed\n");
+		return (NULL);
+	}
+	i = -1;
+	while (++i < d->nbr_lines)
+	{
+		map[i] = malloc(sizeof(unsigned short) * d->len_lines);
+		if (!map[i])
+			return (free_map_i(map, i));
+	}
+	return (map);
+}
